Bug fixed crash when changing R3D scale.
Removed some warnings from code.

+++ mrViewer/src/core/CMedia.h
@@ -1772,7 +1772,7 @@ protected:
                           const unsigned short channels = 4,
                           const image_type::Format format = image_type::kRGBA,
                           const image_type::PixelType pixel_type = image_type::kFloat,
-                          unsigned w = 0, unsigned h = 0);
+                          size_t w = 0, size_t h = 0);
 
 
     unsigned int audio_bytes_per_frame();
